Refactor rack elevation mixin to RackElevationSVG
